Enhanced purgeLegacyLibraries() to be more robust

+++ src/main/java/com/caoccao/javet/interop/JavetLibLoader.java
@@ -63,30 +63,36 @@ public final class JavetLibLoader {
     private static void purgeLegacyLibraries() {
         File tempRootPath = new File(JavetOSUtils.TEMP_DIRECTORY, TEMP_ROOT_NAME);
         try {
-            for (File tempProcessIDPath : tempRootPath.listFiles()) {
-                if (tempProcessIDPath.isDirectory() &&
-                        tempProcessIDPath.lastModified() + MIN_LAST_MODIFIED_GAP_IN_MILLIS < System.currentTimeMillis()) {
-                    try {
-                        boolean isLocked = false;
-                        for (File libFile : tempProcessIDPath.listFiles()) {
-                            if (libFile.delete()) {
-                                LOGGER.logDebug("Deleted {0}.", libFile.getAbsolutePath());
-                            } else {
-                                LOGGER.logDebug("{0} is locked.", libFile.getAbsolutePath());
-                                isLocked = true;
-                                break;
+            if (tempRootPath.exists()) {
+                if (tempRootPath.isDirectory()) {
+                    for (File tempProcessIDPath : tempRootPath.listFiles()) {
+                        if (tempProcessIDPath.isDirectory() &&
+                                tempProcessIDPath.lastModified() + MIN_LAST_MODIFIED_GAP_IN_MILLIS < System.currentTimeMillis()) {
+                            try {
+                                boolean isLocked = false;
+                                for (File libFile : tempProcessIDPath.listFiles()) {
+                                    if (libFile.delete()) {
+                                        LOGGER.logDebug("Deleted {0}.", libFile.getAbsolutePath());
+                                    } else {
+                                        LOGGER.logDebug("{0} is locked.", libFile.getAbsolutePath());
+                                        isLocked = true;
+                                        break;
+                                    }
+                                }
+                                if (!isLocked) {
+                                    if (tempProcessIDPath.delete()) {
+                                        LOGGER.logDebug("Deleted {0}.", tempProcessIDPath.getAbsolutePath());
+                                    } else {
+                                        LOGGER.logDebug("{0} is locked.", tempProcessIDPath.getAbsolutePath());
+                                    }
+                                }
+                            } catch (Throwable t) {
+                                LOGGER.logError(t, "Failed to delete {0}.", tempProcessIDPath.getAbsolutePath());
                             }
                         }
-                        if (!isLocked) {
-                            if (tempProcessIDPath.delete()) {
-                                LOGGER.logDebug("Deleted {0}.", tempProcessIDPath.getAbsolutePath());
-                            } else {
-                                LOGGER.logDebug("{0} is locked.", tempProcessIDPath.getAbsolutePath());
-                            }
-                        }
-                    } catch (Throwable t) {
-                        LOGGER.logError(t, "Failed to delete {0}.", tempProcessIDPath.getAbsolutePath());
                     }
+                } else {
+                    tempRootPath.delete();
                 }
             }
         } catch (Throwable t) {
